remove failed experiment with remapping kj to esc
It's extremely ergonomic, but a massive problem when pasting text that
may contain 'kj'.  Also it was weird that I didn't have it applied
uniformly to all modes in vim, and it wasn't clear if that would have
been a good idea (eg, it interferes badly with movement in normal mode).

Currently working with ctrl-[, which is more difficult but safer.