clean up tmux bootstrap
Checking $TMUX seems to be more reliable than $TERM (this was previously
dependent on a specific value set in tmux.conf).

Use `tmux has-session` to check for the existence of the default session
before trying to attach to it.  Both `has-session` and `attach-session`
print an unwanted error message if the session doesn't exist, but it
should be safer to suppress the output from `has-session` since it does
less.