start tmux first instead of last
This avoids effective double application of .bash_profile and .bashrc in
the common case, so that we don't for example, have duplicate entries in
$PATH.

There are scenarios where .bashrc might run but not .bash_profile (see
https://blog.flowblok.id.au/2013-02/shell-startup-scripts.html for
details) but I don't typically use them and if I did, not starting tmux
is not a big deal.

There might still be an issue for login, non-interactive shells, where
we probably do not want to run tmux, but this commit doesn't change that
situation.  Can be fixed later if encountered.