copy tmux buffer to system pasteboard
So far this seems to be working without the reattach-to-user-namespace
hack.  Not sure why.

This copies automatically when using `y`.  Some guides suggest a
separate C-c binding to make the pbcopy a separate step, but I see no
reason not to do it automatically.

This makes no use of pbpaste because we can't update tmux's buffer
automatically when copying elsewhere in the os.  A C-v binding could do
this for us, but that would create yet another way to paste text.  Given
that we can use cmd-v to paste directly into the terminal anyway, that's
probably good enough.  Since tmux buffer content is now always copied
up to the pasteboard, we could probably also get rid of the p binding
for pasting from tmux.