solarized colours
A neovim update somehow made the default colours really weird so it
makes sense to just store something explicit.

This imports the color scheme for vim and configures it as well as adds
a tweak to the Terminal.app colour scheme to match solarized's
background (otherwise there's a ring of different colour around vim,
partially provided by the terminal, and partially by tmux).

For future reference, to capture the colour in Terminal.app's colour
picker:
- open vim so that the solarized background takes most of the screen
- open the Terminal.app profile
- click Background->Colour&Effects
- click the eyedrop thing
- click the vim background