remove the auto-connection to tmux
This does too many weird things when other things try to start a shell.
Eg, Visual Studio tries to run a shell with a particular command, but
connects to tmux instead and never gets to what it wanted to do.

Can probably do better than just commenting it out, but it will take
some thought.